+++ packages/ui/src/data-table/components/data-table-facet/index.ts
@@ -1,0 +1,1 @@
+export { DataTableFacet } from "./DataTableFacet"